+++ test/LionWeb.Core.Test/Serialization/SerializationLanguageTests.cs
@@ -34,7 +34,8 @@ public class LanguageSerializationTests
     [TestMethod]
     public void shapes_language_with_external_annotations()
     {
-        var serializationChunk = new Serializer(_lionWebVersion).SerializeToChunk([ShapesLanguage.Instance]);
+        var serializationChunk = new SerializerBuilder().WithLionWebVersion(_lionWebVersion).Build()
+            .SerializeToChunk([ShapesLanguage.Instance]);
 
         var redeserialized = ILanguageDeserializerExtensions.Deserialize(serializationChunk);
         Language redeserializedShapes = redeserialized.Cast<INode>().OfType<Language>().First();
@@ -55,7 +56,8 @@ public class LanguageSerializationTests
         annInst.Set(age, 23);
         redeserializedCircle.AddAnnotations([annInst]);
 
-        var serializationChunk2 = new Serializer(_lionWebVersion).SerializeToChunk([redeserializedShapes]);
+        var serializationChunk2 = new SerializerBuilder().WithLionWebVersion(_lionWebVersion).Build()
+            .SerializeToChunk([redeserializedShapes]);
         Console.WriteLine(JsonUtils.WriteJsonToString(serializationChunk2));
         var redeserialized2 = ILanguageDeserializerExtensions.Deserialize(serializationChunk2, [language]);
         Language redeserializedShapes2 = redeserialized2.Cast<INode>().OfType<Language>().First();
@@ -70,11 +72,14 @@ public class LanguageSerializationTests
     [TestMethod]
     public void shapes_language_together_with_annotations()
     {
-        var serializationChunk = new Serializer(_lionWebVersion).SerializeToChunk([ShapesLanguage.Instance]);
+        var serializationChunk = new SerializerBuilder().WithLionWebVersion(_lionWebVersion).Build()
+            .SerializeToChunk([ShapesLanguage.Instance]);
 
-        var redeserialized =
-            new LanguageDeserializer(_lionWebVersion, compressedIdConfig: new(KeepOriginal: true))
-                .Deserialize(serializationChunk);
+        var redeserialized = new LanguageDeserializerBuilder()
+            .WithLionWebVersion(_lionWebVersion)
+            .WithCompressedIds(new(KeepOriginal: true))
+            .Build()
+            .Deserialize(serializationChunk);
         Language redeserializedShapes = redeserialized.Cast<INode>().OfType<Language>().First();
 
         var language =
@@ -101,11 +106,14 @@ public class LanguageSerializationTests
 
         redeserializedCircle.AddAnnotations([annInst]);
 
-        var serializationChunk2 = new Serializer(_lionWebVersion).SerializeToChunk([redeserializedShapes, language]);
+        var serializationChunk2 = new SerializerBuilder().WithLionWebVersion(_lionWebVersion).Build()
+            .SerializeToChunk([redeserializedShapes, language]);
         Console.WriteLine(JsonUtils.WriteJsonToString(serializationChunk2));
-        var redeserialized2 =
-            new LanguageDeserializer(_lionWebVersion, compressedIdConfig: new(KeepOriginal: true)).Deserialize(
-                serializationChunk2);
+        var redeserialized2 = new LanguageDeserializerBuilder()
+            .WithLionWebVersion(_lionWebVersion)
+            .WithCompressedIds(new(KeepOriginal: true))
+            .Build()
+            .Deserialize(serializationChunk2);
         Language redeserializedShapes2 = redeserialized2.Cast<INode>().OfType<Language>()
             .First(l => l.Key == ShapesLanguage.Instance.Key);
         DynamicClassifier redeserializedCircle2 =
@@ -119,11 +127,16 @@ public class LanguageSerializationTests
     [TestMethod]
     public void LionCore()
     {
-        var serializationChunk = new Serializer(_lionWebVersion).SerializeToChunk([_m3]);
+        var serializationChunk =
+            new SerializerBuilder().WithLionWebVersion(_lionWebVersion).Build().SerializeToChunk([_m3]);
         Console.WriteLine(JsonUtils.WriteJsonToString(serializationChunk));
 
         // Just run the deserializer for now (without really checking anything), to see whether it crashes or not:
-        var deserializer = new LanguageDeserializer(_lionWebVersion, new SkipDeserializationHandler());
+        var deserializer = new LanguageDeserializerBuilder()
+            .WithLionWebVersion(_lionWebVersion)
+            .WithHandler(new SkipDeserializationHandler())
+            .Build();
+
         foreach (var serializedNode in serializationChunk.Nodes)
         {
             deserializer.Process(serializedNode);
@@ -147,12 +160,18 @@ public class LanguageSerializationTests
             LionWebVersions.v2024_1.LionCore,
             LionWebVersions.v2024_1.BuiltIns
         ];
-        var chunk =
-            new Serializer(LionWebVersions.v2024_1_Compatible) { CompressedIdConfig = new(KeepOriginal: true) }
-                .SerializeToChunk(input);
 
-        var deserializer =
-            new LanguageDeserializer(LionWebVersions.v2024_1_Compatible, new SkipDeserializationHandler());
+        var serializer = new SerializerBuilder()
+            .WithLionWebVersion(LionWebVersions.v2024_1_Compatible)
+            .WithCompressedIds(new(KeepOriginal: true))
+            .Build();
+
+        var chunk = serializer.SerializeToChunk(input);
+
+        var deserializer = new LanguageDeserializerBuilder()
+            .WithLionWebVersion(LionWebVersions.v2024_1_Compatible)
+            .WithHandler(new SkipDeserializationHandler())
+            .Build();
 
         var actual = deserializer.Deserialize(chunk).Cast<Language>().ToList();
 
@@ -179,7 +198,8 @@ public class LanguageSerializationTests
     [TestMethod]
     public void shapes_language()
     {
-        var serializationChunk = new Serializer(_lionWebVersion).SerializeToChunk([ShapesLanguage.Instance]);
+        var serializationChunk = new SerializerBuilder().WithLionWebVersion(_lionWebVersion).Build()
+            .SerializeToChunk([ShapesLanguage.Instance]);
         Console.WriteLine(JsonUtils.WriteJsonToString(serializationChunk));
 
         var redeserialized = ILanguageDeserializerExtensions.Deserialize(serializationChunk);