+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres _loosely_ to [Semantic Versioning](https://semver.org/s
 ## [Unreleased]
 
 ### Added
+* Introduced `SerializerBuilder`.
+* Introduced `LanguageDeserializerBuilder`.
 ### Fixed
 ### Changed
 ### Removed